Use approx 512Mb per thread for splitting.

+++ assembler/src/hammer/kmer_data.cpp
@@ -150,7 +150,7 @@ path::files_t HammerKMerSplitter::Split(size_t num_files) {
     KMerBuffer &entry = tmp_entries[i];
     entry.resize(num_files);
     for (unsigned j = 0; j < num_files; ++j) {
-      entry[j].reserve((long unsigned)(1.1 * (double)cell_size));
+      entry[j].reserve((size_t)(1.1 * (double)cell_size));
     }
   }
